修复douOrder错误

+++ service/dou/src/main/java/cn/qihangerp/module/open/dou/domain/DouOrder.java
@@ -38,7 +38,7 @@ public class DouOrder implements Serializable {
     /**
      * 订单状态1 待确认/待支付（订单创建完毕）105 已支付 2 备货中 101 部分发货 3 已发货（全部发货）4 已取消5 已完成（已收货）
      */
-    private Long orderStatus;
+    private Integer orderStatus;
 
     /**
      * 订单状态描述
